+++ packages/Flow/src/package-test.ts
@@ -43,6 +43,7 @@ import './tests/output-views-tests';
 import './tests/uploaded-file-tests';
 import './tests/toolbox-tabs-tests';
 import './tests/input-value-tests';
+import './tests/property-input-tests';
 import './tests/slot-visibility-tests';
 
 export let _package = new DG.Package();